requirements.txt
astroid==1.6.5
certifi==2018.10.15
chardet==3.0.4
Django==2.0.8
django-anymail==5.0
django-ckeditor==5.6.1
django-extensions==2.0.7
django-js-asset==1.1.0
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.4.1
pkg-resources==0.0.0
PyQRCode==1.2.1
python-dateutil==2.8.0
pytz==2018.4
requests==2.20.1
six==1.11.0
urllib3==1.24.1
wrapt==1.10.11
